requirements.txt
Django>=5,<6
 djangorestframework>=3.15,<4
 pytest>=8,<9
 pytest-django>=4.9,<5
